requirements.txt
# Core dependencies
kubernetes>=28.1.0
prometheus-api-client>=0.5.3
pandas>=2.0.0
numpy>=1.24.0

# ML libraries
prophet>=1.1.4
scikit-learn>=1.3.0
statsmodels>=0.14.0

# CLI and UI
rich>=13.0.0
click>=8.1.0
pyyaml>=6.0
tabulate>=0.9.0

# HTTP and networking
requests>=2.31.0
urllib3>=2.0.0

# Data processing
matplotlib>=3.7.0
seaborn>=0.12.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: for advanced time series analysis
arch>=6.2.0

